tests: moved signal.notch test to new module
